Taking out the benchmark stuff
