+++ main/src/ca/uwaterloo/flix/Main.scala
@@ -19,7 +19,7 @@ package ca.uwaterloo.flix
 import java.io.File
 
 import ca.uwaterloo.flix.api.{Flix, MatchException, RuleException, SwitchException, UserException}
-import ca.uwaterloo.flix.runtime.Benchmark.FlixBenchmark
+import ca.uwaterloo.flix.runtime.Benchmark.Benchmarker
 import ca.uwaterloo.flix.runtime._
 import ca.uwaterloo.flix.util._
 import ca.uwaterloo.flix.util.vt.VirtualString.{Code, Line, NewLine}
@@ -146,7 +146,7 @@ object Main {
           }
 
           if (cmdOpts.benchmark) {
-            FlixBenchmark.benchmark(model)
+            Benchmarker.benchmark(model)
           }
 
           if (cmdOpts.test) {
